+++ src/fast-clash/clashFilter.cpp
@@ -708,8 +708,9 @@ void clashFilter::approximateInteractions(Matrix transMat, int nodeS, int nodeM,
   double sumRad = staticAtomsOctree[nodeS].cr + movingAtomsOctree[nodeM].cr;
   double sumRad2 = sumRad * sumRad;
 
-  Vector oldPos(movingAtomsOctree[nodeM].cx, movingAtomsOctree[nodeM].cy,
-                movingAtomsOctree[nodeM].cz, 1.0);
+  Vector oldPos((float)(movingAtomsOctree[nodeM].cx),
+                (float)(movingAtomsOctree[nodeM].cy),
+                (float)(movingAtomsOctree[nodeM].cz), 1.0);
   Vector newPos = transMat * oldPos;
   double cxM = newPos[0], cyM = newPos[1], czM = newPos[2];
 
@@ -733,8 +734,8 @@ void clashFilter::approximateInteractions(Matrix transMat, int nodeS, int nodeM,
     if (staticAtomsOctree[nodeS].leaf && movingAtomsOctree[nodeM].leaf) {
       for (int i = movingAtomsOctree[nodeM].atomsStartID;
            i <= movingAtomsOctree[nodeM].atomsEndID; i++) {
-        Vector oldPos(movingAtoms[i].x, movingAtoms[i].y, movingAtoms[i].z,
-                      1.0);
+        Vector oldPos((float)(movingAtoms[i].x), (float)(movingAtoms[i].y),
+                      (float)(movingAtoms[i].z), 1.0);
         Vector newPos = transMat * oldPos;
         double xM = newPos[0], yM = newPos[1], zM = newPos[2];
 
@@ -903,7 +904,8 @@ bool clashFilter::computeInteractionsNaively(Matrix transMat, int *nClashes,
 
   int numClose = 0;
   for (int i = 0; i < nMovingAtoms; i++) {
-    Vector oldPos(movingAtoms[i].x, movingAtoms[i].y, movingAtoms[i].z, 1.0);
+    Vector oldPos((float)(movingAtoms[i].x), (float)(movingAtoms[i].y),
+                  (float)(movingAtoms[i].z), 1.0);
     Vector newPos = transMat * oldPos;
 
     double xM = newPos[0], yM = newPos[1], zM = newPos[2];
